requirements.txt
Flask>=3.1.1
python-json-logger>=3.2.1
requests>=2.32.3
python-dotenv>=1.0.1
gunicorn>=23.0.0
Werkzeug>=3.1.3
ddtrace>=3.18.1

